+++ bankingapp/src/main/java/com/synergisticit/bankingapp/repository/RoleRepository.java
@@ -2,9 +2,11 @@ package com.synergisticit.bankingapp.repository;
 
 import com.synergisticit.bankingapp.domain.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Optional<Role> findByRoleName(String roleName);
 }
